feat: architecture views become a registry doc type on a Named identity (ADR 0036)
Adds Identity::Named (dir/slug.md, no number) and registers the view doc
type in docs/architecture/: scaffolded by 'new view --kind' (closed
C4/arc42 vocabulary, validated), listed by a generated index sorted in
zoom order (kind rank, then filename), canonically checked, and covered
by the docs-handwrite hook. Views carry no status and are never
superseded — living documents updated in place.

Registers the model-based distill direction (Structurizr/LikeC4-style
views generated from the ADR 0031/0032 graph) as research 0004, not a
decision. Splits oversized modules into sibling test files to honor the
300-line ratchet (new/fill, doc_type, brief, index).
