+++ living-docs-core/src/paths.rs
@@ -7,7 +7,7 @@ use crate::doc_type::{self, Identity};
 pub fn dir_for(token: &str) -> Option<&'static str> {
     match doc_type::spec_for(token)?.identity {
         Identity::Numbered { dir } => Some(dir),
-        Identity::Singleton { .. } => None,
+        Identity::Singleton { .. } | Identity::Named { .. } => None,
     }
 }
