+++ scripts/file-size-baseline.txt
@@ -14,12 +14,9 @@
 304 living-docs-core/src/check/graph.rs
 332 living-docs-core/src/check/mermaid.rs
 380 living-docs-core/src/check/records.rs
-523 living-docs-core/src/commands/brief.rs
 415 living-docs-core/src/commands/export.rs
-881 living-docs-core/src/commands/index.rs
+454 living-docs-core/src/commands/index.rs
 1023 living-docs-core/src/commands/leak_gate.rs
-759 living-docs-core/src/commands/new.rs
-478 living-docs-core/src/doc_type.rs
 419 living-docs-core/src/pii/brazil.rs
 918 living-docs-core/src/pii/contextual.rs
 534 living-docs-core/src/pii/europe.rs
