+++ skills/living-docs/templates/architecture-index.md
@@ -1,44 +1,18 @@
 <!-- OKF reserved index.md (§6): a directory listing — NO frontmatter. The view files
-     it links ARE concepts and each carries `type: Architecture View` frontmatter. -->
+     it links ARE concepts (`type: Architecture View`), scaffolded by
+     `living-docs new view "<title>" --kind <kind>` and listed by `living-docs index`
+     in C4/arc42 zoom order (ADR 0036). Everything above the first row is preamble the
+     generator preserves; the rows below it are generator-owned — never hand-edit them. -->
 
 # Architecture
 
-Living architecture of the project: how the components fit together, expressed as
-Mermaid diagrams that must always match the code. This index is the entry point; each
-view file owns one diagram/concern. Update the relevant view in the same change as any
-structural code change (see the maintenance rule in the project guide).
+Living architecture of the project, one view per concern, sequenced outside-in:
+context, containers, components, then runtime behavior (flows, sequences, states),
+data, and deployment. Each view names its drift instrument; update the relevant view
+in the same change as any structural code change (see `rules/maintenance-invariant.md`).
 
-## Views
-
-| File | View | Mermaid type |
-|---|---|---|
-| [context.md](context.md)         | High-level components & external world | `flowchart` |
-| [data-model.md](data-model.md)   | Entities & relationships               | `erDiagram` |
-| [modules.md](modules.md)         | Module layout & dependencies           | `flowchart` |
-| [<flow>.md](flow.md)             | <Key process / data flow>              | `flowchart` |
-| [<request>.md](request.md)       | <Tool-calling / request round trip>    | `sequenceDiagram` |
-
-<!-- Add a row whenever a new view file is created. -->
-
-## Example view file skeleton
-
-A view file is an OKF concept: it opens with frontmatter, then an H1 title, one sentence
-of orientation, then the Mermaid block:
-
-````markdown
----
-type: Architecture View
-title: <View name>
-description: <One sentence — what this view shows.>
-timestamp: <ISO 8601 datetime>
----
-
-# <View name>
-
-<One sentence: what this view shows and when to consult it.>
-
-```mermaid
-flowchart LR
-    A[Module A] --> B[Module B]
-```
-````
+* [Context](context.md) - context
+* [Backends](backends.md) - container
+* [Module layout](modules.md) - component
+* [Request round trip](request-round-trip.md) - sequence
+* [Data model](data-model.md) - data-model
